Fix validForm function location
